requirements.txt
Django>=5.0
Pillow>=10.0.0
python-decouple>=3.8
django-environ>=0.11.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.7
whitenoise>=6.6.0
django-cors-headers>=4.3.0
